fix: stop two field-re-listing node rebuilds and a float4 type drift
Review round 3 found three concrete issues:

- ProjectionPushdownRule._prune_columns rebuilt a Filter with a raw constructor
  (Filter(input=..., predicate=...)) while its sibling arms (Projection, Join,
  Aggregate, Limit, Sort) all use model_copy - the field-re-listing pattern
  CLAUDE.md forbids, which drops a field if Filter ever gains one. Use model_copy.

- parser._guard_aggregate_args (the agg(x) FILTER (WHERE ...) rewrite) rebuilt a
  FunctionCall by re-listing function_name/args/is_aggregate/distinct, silently
  DROPPING within_group_key/within_group_desc - so a FILTERed ordered-set
  aggregate (PERCENTILE_CONT(..) WITHIN GROUP (..) FILTER (..)) lost its ordering.
  Rebuild via model_copy, preserving every field while replacing only the args.

- Postgres _OID_TO_ARROW mapped float4 (OID 700) to float64, but the catalog
  maps "real" -> DataType.FLOAT -> float32, so a real column was typed two ways
  (and the psycopg2 path also disagreed with ADBC). float4 is 32-bit; map it to
  float32 so fetch, catalog, and ADBC agree.

Guard tests: a FILTERed ordered-set aggregate keeps its within-group key; a plain
FILTERed aggregate still guards its arg; the float OIDs agree with the catalog's
native-name -> DataType -> Arrow path. Full suite 1053 pass.

+++ federated_query/datasources/postgresql.py
@@ -502,7 +502,7 @@ class PostgreSQLDataSource(DataSource):
         20: pa.int64(),  # int8
         21: pa.int64(),  # int2
         23: pa.int64(),  # int4
-        700: pa.float64(),  # float4
+        700: pa.float32(),  # float4 (32-bit; matches DataType.FLOAT and ADBC)
         701: pa.float64(),  # float8
         1700: pa.float64(),  # numeric
         25: pa.string(),  # text
